Complete src layout: extract state, blocklist, scheduler modules; split tests
state.py:     load_state, save_state, write_attention, _had_attention, STATE_VERSION
blocklist.py: resolve_source, parse_*_blocklist, channel_to_url, check_removals
scheduler.py: _parse_schedule_hours, _format_hours, _find_installed_binary,
              _schedule_macos, _schedule_linux, schedule_cmd

__init__.py slimmed to main() + re-exports of all public names from sub-modules.
browser.py unchanged (already separated in previous pass).

Sub-modules use _pkg() late-import helpers to resolve live-patched attributes
(STATE_FILE, fetch_remote, _schedule_linux, etc.) from yt_dont_recommend,
preserving all existing patch("yt_dont_recommend.X") test targets.

Test file split by module:
  test_state.py, test_blocklist.py, test_scheduler.py, test_browser.py
tests/test_yt_dont_recommend.py removed.

All 109 tests pass. Zero behaviour changes.
